Added sequential_merge method to address_collinearity_with_vif.
This method will iteratively merge high collinearity columns with the column(s) they are most highly correlated with until the vif threshold is met. This is an alternative to the regular sequential method which drops columns.

The merge method is controllable via the merge_function argument, which defaults to taking the max of all the columns to be merged.

PiperOrigin-RevId: 569111147